// apps/arabic-mandi/src/app/Core/admin/generate-admin-types.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type BooleanFieldComparison = {
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
};
// ...
export type NumberFieldComparison = {
  between?: InputMaybe<NumberFieldComparisonBetween>;
  eq?: InputMaybe<Scalars['Float']>;
  gt?: InputMaybe<Scalars['Float']>;
  gte?: InputMaybe<Scalars['Float']>;
  in?: InputMaybe<Array<Scalars['Float']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  lt?: InputMaybe<Scalars['Float']>;
  lte?: InputMaybe<Scalars['Float']>;
  neq?: InputMaybe<Scalars['Float']>;
  notBetween?: InputMaybe<NumberFieldComparisonBetween>;
  notIn?: InputMaybe<Array<Scalars['Float']>>;
};
// ...
export type OfferFilter = {
  and?: InputMaybe<Array<OfferFilter>>;
  createdby?: InputMaybe<OfferFilterUserFilter>;
  discount?: InputMaybe<NumberFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<OfferFilter>>;
  price?: InputMaybe<NumberFieldComparison>;
  status?: InputMaybe<BooleanFieldComparison>;
  totalPrice?: InputMaybe<NumberFieldComparison>;
  updatedby?: InputMaybe<OfferFilterUserFilter>;
};

export type OfferFilterUserFilter = {
  and?: InputMaybe<Array<OfferFilterUserFilter>>;
  id?: InputMaybe<NumberFieldComparison>;
  or?: InputMaybe<Array<OfferFilterUserFilter>>;
  role?: InputMaybe<StringFieldComparison>;
  status?: InputMaybe<BooleanFieldComparison>;
};
// ...
export type OfferSort = {
  direction: SortDirection;
  field: OfferSortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum OfferSortFields {
  Discount = 'discount',
  Name = 'name',
  Price = 'price',
  Status = 'status',
  TotalPrice = 'totalPrice'
}
// ...
export enum SortDirection {
  Asc = 'ASC',
  Desc = 'DESC'
}

/** Sort Nulls Options */
export enum SortNulls {
  NullsFirst = 'NULLS_FIRST',
  NullsLast = 'NULLS_LAST'
}

export type StringFieldComparison = {
  eq?: InputMaybe<Scalars['String']>;
  gt?: InputMaybe<Scalars['String']>;
  gte?: InputMaybe<Scalars['String']>;
  iLike?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<Scalars['String']>>;
  is?: InputMaybe<Scalars['Boolean']>;
  isNot?: InputMaybe<Scalars['Boolean']>;
  like?: InputMaybe<Scalars['String']>;
  lt?: InputMaybe<Scalars['String']>;
  lte?: InputMaybe<Scalars['String']>;
  neq?: InputMaybe<Scalars['String']>;
  notILike?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<Scalars['String']>>;
  notLike?: InputMaybe<Scalars['String']>;
};
// ...
export type UpdateOfferInput = {
  discount?: InputMaybe<Scalars['Float']>;
  items?: InputMaybe<Array<OfferItemInput>>;
  name?: InputMaybe<Scalars['String']>;
  price?: InputMaybe<Scalars['Float']>;
  status?: InputMaybe<Scalars['Boolean']>;
  totalPrice?: InputMaybe<Scalars['Float']>;
  updatedby?: InputMaybe<UserInput>;
};
// ...
export type UpdateOneOfferInput = {
  /** The id of the record to update */
  id: Scalars['ID'];
  /** The update to apply. */
  update: UpdateOfferInput;
};
// ...
export type UpdateOneOfferMutationVariables = Exact<{
  input: UpdateOneOfferInput;
}>;


export type UpdateOneOfferMutation = { __typename?: 'Mutation', updateOneOffer: { __typename?: 'Offer', id: string } };

export type CreateManyOfferMutationVariables = Exact<{
  input: CreateManyOffersInput;
}>;


export type CreateManyOfferMutation = { __typename?: 'Mutation', createManyOffers: Array<{ __typename?: 'Offer', name: string, id: string }> };

export type GetAlloffersQueryVariables = Exact<{
  filter?: OfferFilter;
  sorting?: Array<OfferSort> | OfferSort;
}>;